requirements.txt
fastapi
uvicorn
playwright
pytrends
pandas
feedparser
requests
pyyaml
google-genai
python-dotenv
beautifulsoup4
httpx
urllib3<2
supabase
openpyxl
instagrapi
curl_cffi
cloudscraper
